+++ pkg/analyzer/lib/src/dart/resolver/function_reference_resolver.dart
@@ -211,7 +211,7 @@ class FunctionReferenceResolver {
           nameNode.name,
           element.kind.displayName,
           enclosingElement.name!,
-          enclosingElement is ClassElement && enclosingElement.isMixin
+          enclosingElement is MixinElement
               ? 'mixin'
               : enclosingElement.kind.displayName,
         ],