+++ pkg/analyzer/lib/src/dart/resolver/method_invocation_resolver.dart
@@ -256,7 +256,7 @@ class MethodInvocationResolver with ScopeHelpers {
           nameNode.name,
           element.kind.displayName,
           enclosingElement.name!,
-          enclosingElement is ClassElement && enclosingElement.isMixin
+          enclosingElement is MixinElement
               ? 'mixin'
               : enclosingElement.kind.displayName,
         ],